Moved PSstatSender integration tests to separate test binary
