+++ test/Makefile.am
@@ -4,7 +4,7 @@ AM_CPPFLAGS = -I$(top_srcdir)/include -I$(top_srcdir)/3rdparty @PS_FLAGS@
 LDADD = ../src/libchimbuko.la -lgtest -lstdc++fs 
 
 testdir = $(prefix)/test
-test_PROGRAMS = mainSimple mainNet mainAd
+test_PROGRAMS = mainSimple mainNet mainAd mainStatSender
 
 mainSimple_SOURCES = mainSimple.cpp test_utils.cpp test_param.cpp
 mainSimple_LDADD = $(LDADD)
@@ -15,7 +15,12 @@ mainNet_LDADD = $(LDADD)
 mainAd_SOURCES = mainAD.cpp test_ad.cpp
 mainAd_LDADD = $(LDADD)
 
-test_DATA = run_all.sh run_ad.sh run_net.sh
+mainStatSender_SOURCES = test_stat_sender.cpp
+mainStatSender_LDADD = $(LDADD)
+
+
+
+test_DATA = run_all.sh run_ad.sh run_net.sh run_stat_sender.sh
 
 if ENABLE_PROVDB
 test_PROGRAMS += mainProvDBclient mainADwithProvDB
